+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 **Features**
 - Added new config parameter "max-fuse-threads" under "libfuse" config to control max threads allowed at libfuse layer.
 - Added new config parameter 'refresh-sec' in 'file-cache'. When file-cache-timeout is set to a large value, this field can control when to refresh the file if file in container has changed.
+- Added FUSE option `dirrect_io` to bypass the kernel cache and perform direct I/O operations.
 
 **Bug Fixes**
 - [#1116](https://github.com/Azure/azure-storage-fuse/issues/1116)] Relative path for tmp-cache is resulting into file read-write failure.